backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pandas==2.1.3
sqlalchemy==2.0.23
pyodbc==5.0.1
openai==1.12.0
matplotlib==3.8.2
seaborn==0.13.0
python-multipart==0.0.6
pydantic==2.5.2
numpy==1.26.2
httpx==0.24.1
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
openpyxl==3.1.2 